[REFACTOR] CertiK Audit Recommendation - Exhibit 9
Description:
The constant variable “MAX_UINT” is meant to represent the maximum value a “uint256” can
hold. In the aforementioned line, the expression of that value is stated as “2 ^ 256 - 1”.

Recommendations:
This expression is incorrect as the “^” operand in Solidity is a “XOR” operation (bitwise
exclusive-or) rather than a “power-of” operation. Thus, the result of “2 ^ 256 - 1” evaluates to
“253”. A “power-of” operation is expressed by two consecutive asterisks “**”, consult Exhibit 6
for the correct statement(s).