[REFACTOR] Certik Audit Recommendation - Exhibit 16
Description:
The “require” statements provide no “reason” string for when they fail.

Recommendations:
A corresponding error code should be provided, as done with all other “require” statements, to
ease the debugging of smart contracts utilizing “NftStaking” as a base. To note, error code “9” is
unused in the sequential list of codes mentioned in the comment block before the contract
declaration.