[REFACTOR] Certik Audit Recommendation - Exhibit 30
Description:
At the specified lines the “catch” block of a low-level EVM error can be observed. However, the
exact same statement that caused the error to be thrown is executed.

Recommendations:
The error should be handled differently or not caught at all if bubbling up the error in the call
chain is desired. A corresponding error code should also be devised.