[CHANGED] Certik Audit - Exhibit 2 (round 3)
[INFORMATIONAL] Description:
The code currently duplicates the exact same statements across two different types of “catch” blocks.

Recommendations:
As the “reason” and “lowLevelData” variables are not utilized, it is possible to replace both “catch” clauses with a single empty clause like so “catch { .. }“ thus reducing the number of statements, logical jumps and subsequently bytecode size.